got flow passing and ready to drive this out with tests
